Rewrite World class

+++ src/main/java/org/egordorichev/lasttry/entity/EnemyID.java
@@ -1,8 +1,8 @@
 package org.egordorichev.lasttry.entity;
 
 public class EnemyID {
-	public static final int none = 0;
-	public static final int greenSlime = 1;
-	public static final int blueSlime = 2;
-	public static final int eyeOfCthulhu = 3;
+	public static final short none = 0;
+	public static final short greenSlime = 1;
+	public static final short blueSlime = 2;
+	public static final short eyeOfCthulhu = 3;
 }
