+++ src/main/java/org/egordorichev/lasttry/item/Mushroom.java
@@ -3,7 +3,7 @@ package org.egordorichev.lasttry.item;
 import org.newdawn.slick.Image;
 
 public class Mushroom extends Item { // TODO
-	public Mushroom(int id, String name, Image texture) {
+	public Mushroom(short id, String name, Image texture) {
 		super(id, name, texture);
 	}
