+++ src/main/java/org/egordorichev/lasttry/item/SlippyBlock.java
@@ -1,10 +1,10 @@
 package org.egordorichev.lasttry.item;
 
-import org.egordorichev.lasttry.item.tiles.BlockGround;
+import org.egordorichev.lasttry.item.block.BlockGround;
 import org.newdawn.slick.Image;
 
 public class SlippyBlock extends BlockGround { // TODO
-	public SlippyBlock(int id, String name, Image texture, Image tiles) {
+	public SlippyBlock(short id, String name, Image texture, Image tiles) {
 		super(id, name, texture, tiles);
 	}
 }
