+++ src/main/java/org/egordorichev/lasttry/item/ThornBlock.java
@@ -3,7 +3,7 @@ package org.egordorichev.lasttry.item;
 import org.newdawn.slick.Image;
 
 public class ThornBlock extends Item { // TODO
-	public ThornBlock(int id, String  name, Image texture) {
+	public ThornBlock(short id, String  name, Image texture) {
 		super(id, name, texture);
 	}
 }
