+++ src/main/java/org/egordorichev/lasttry/item/items/Accessory.java
@@ -4,7 +4,7 @@ import org.egordorichev.lasttry.item.Item;
 import org.newdawn.slick.Image;
 
 public class Accessory extends Item {
-	public Accessory(int id, String name, Image texture) {
+	public Accessory(short id, String name, Image texture) {
 		super(id, name, texture);
 	}
 }
