+++ src/main/java/org/egordorichev/lasttry/item/items/Coin.java
@@ -4,7 +4,7 @@ import org.egordorichev.lasttry.item.Item;
 import org.newdawn.slick.Image;
 
 public class Coin extends Item {
-	public Coin(int id, String name, Image texture) {
+	public Coin(short id, String name, Image texture) {
 		super(id, name, texture);
 	}
