+++ src/main/java/org/egordorichev/lasttry/item/items/Dye.java
@@ -6,7 +6,7 @@ import org.newdawn.slick.*;
 public class Dye extends Item {
 	protected Color color;
 	
-	public Dye(int id, String name, Color color, Image texture) {
+	public Dye(short id, String name, Color color, Image texture) {
 		super(id, name, texture);
 	
 		this.color = color;
