+++ src/main/java/org/egordorichev/lasttry/item/items/Pickaxe.java
@@ -4,11 +4,11 @@ import org.egordorichev.lasttry.item.Rarity;
 import org.newdawn.slick.Image;
 
 public class Pickaxe extends Tool {
-	public Pickaxe(int id, String name, Rarity rarity, float baseDamage, int pickaxePower, int useSpeed, Image texture) {
+	public Pickaxe(short id, String name, Rarity rarity, float baseDamage, int pickaxePower, int useSpeed, Image texture) {
 		super(id, name, rarity, baseDamage, pickaxePower, 0, 0, useSpeed, texture);
 	}
 
-	public Pickaxe(int id, String name, float baseDamage, int pickaxePower, int useSpeed, Image texture) {
+	public Pickaxe(short id, String name, float baseDamage, int pickaxePower, int useSpeed, Image texture) {
 		this(id, name, Rarity.WHITE, baseDamage, pickaxePower, useSpeed, texture);
 	}
